refactor(state): split database access into modules by Zebra types
Also split the genesis block check from the genesis note commitment trees.
